Aktualisieren der Bash Skripte

+++ source/cleanup.bash
@@ -5,7 +5,7 @@ IFS=$'\n\t' # Strict Internal Field Separator
 
 for file in *; do
   fileextension=${file##*.}
-  if [[ -f $file && ($fileextension != "tex" && $fileextension != "dtx" && $fileextension != "ins" && $fileextension != ".bib"  && $fileextension != "bash")]]; then
+  if [[ -f $file && ($fileextension != "tex" && $fileextension != "biblio" && $fileextension != "dtx" && $fileextension != "ins" && $fileextension != ".bib"  && $fileextension != "bash")]]; then
     #echo $file
     rm $file
   fi
